Fixes for gdscript syntax checking (#7)
* Fixed the typo in the folder name configrations

Now the name is configurations. All the code
referencing the folder has also been updated.

* Updated deprecated variables

* Demonstration of current diagnostics issues

All the lines that get a complaint from diagnostics are marked
with #in 0.2.2 and a short explanation whether the complaint is valid
or not.

* Enabled syntax highlighting for comments

Now comments have a different colour compared to variables. Found it
unclear which file controls syntax so edited both
GDScript.full.tmLanguage.json and GDScript.tmLanguage.json. In addition
renamed test_files directory to be clearer and started working on
diagnostic.ts to fix false positives in syntax highglighting.

* Fixed checking comments for syntax

Now comments are not checked for syntax. This was achieved by excluding
comments from being parsed. In addition made it so that syntax check no
longer complains about empty lines having wrong indentation.

* Wrong committer details on previous commits

This commit should have the correct details.

* Added a comment explaining how nextline is parsed

* Reverted changes in symbolparser.ts

Also made minor additions to test_files.