Properly kill headless LSP process on Linux
On Linux, child processes are allowed to persist when their parent is terminated.

Currently, headless LSP processes are spawned with `shell: true` option, which
means a shell process is launched first, and its PID is returned. Calling
`process.kill()` on that does kill the shell process, but the Godot Engine
instance launched by the shell keeps running.

Instead, kill the process group of which the shell is the leader (guaranteed by
the `detached: true` option). In Linux, process group IDs are equal to PIDs
of their leader processes, and negative PIDs are used to send a signal to a group.
Thus, negating the PID in `process.kill()` call is enough.