Fix Inherits in documentation
It seems the extended inheritence chain previously being parsed is no
longer available, which causes the "Inherits: " information to be empty.
Inspecting the symbol information available, it seems there is no way to
recover the inheritence chain without additional queries or modifying
LSP response, so a quick fix to display the correct information that we
do have is an incremental improvement.