Change scene dropEdit from $-import to @onready template
The right-click action, that drops the node-path into your clipboard
already provides close to the same utility of the drag-n-drop feature -
only missing a `$`.
To complement this and to add another QoL change, I refactored the
drop-edit functionality to include a full `@onready` template. This will
now create a variable name based on the node-name and provides the full
convention, inspired by the same feature in the godot editor: https://github.com/godotengine/godot-proposals/issues/4482