Use node-label for generating import statements
Originally, we used and split the path of a node to get its name. This
ran into issues with the root-node of a scene though, as a root node has
no path. This logic has been replaced by directly getting the name of
the node from its `label` attribute, which is now passed directly to the
drag handler.